Merchant S C

+++ Application/Services/OrderService.cs
@@ -6,6 +6,7 @@ using Application.Interfaces.ApplicationServices;
 using Application.Interfaces.Repositories;
 using AutoMapper;
 using Domain.Entities;
+using Domain.Enums;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -19,12 +20,20 @@ namespace Application.Services
         private readonly IUnitOfWork _unitOfWork;
         private readonly IMapper _mapper;
         private readonly IPaginationRepository<Order> _repository;
+        private readonly IGenericRepository<City> _cityRepository;
+        private readonly IGenericRepository<Settings> _settingsRepository;
+        private readonly IGenericRepository<Merchant> _merchantRepository;
+        private readonly IGenericRepository<Shipping> _shippingRepository;
 
         public OrderService(IUnitOfWork unitOfWork, IMapper mapper)
         {
             _unitOfWork = unitOfWork;
             _mapper = mapper;
             _repository = _unitOfWork.GetPaginationRepository<Order>();
+            _cityRepository = _unitOfWork.GetGenericRepository<City>();
+            _settingsRepository = _unitOfWork.GetGenericRepository<Settings>();
+            _merchantRepository = _unitOfWork.GetGenericRepository<Merchant>();
+            _shippingRepository = _unitOfWork.GetGenericRepository<Shipping>();
         }
 
         public async Task<List<DisplayOrderDTO>> GetAllObjects()
@@ -63,25 +72,80 @@ namespace Application.Services
             return _mapper.Map<DisplayOrderDTO>(order);
         }
 
-        public Task<bool> InsertObject(InsertOrderDTO orderDTO)
+        public async Task<ModificationResultDTO> InsertObject(InsertOrderDTO orderDTO)
         {
             var order = _mapper.Map<Order>(orderDTO);
-            var result =  _repository.Add(order); 
-            return Task.FromResult(result);
+            
+            order.ShippingCost = await CalculateShipmentCost(order);
+
+            var result =  _repository.Add(order);
+
+            if (result == false)
+            {
+                return new ModificationResultDTO()
+                {
+                    Succeeded = false,
+                    Message = "Error inserting the order"
+                };
+            }
+
+            return new ModificationResultDTO()
+            {
+                Succeeded = true
+            };
         }
 
-        public Task<bool> UpdateObject(UpdateOrderDTO orderDTO)
+        public async Task<ModificationResultDTO> UpdateObject(UpdateOrderDTO orderDTO)
         {
             var order = _mapper.Map<Order>(orderDTO);
-            var result = _repository.Edit(order); 
-            return Task.FromResult(result);
+
+            order.ShippingCost = await CalculateShipmentCost(order);
+
+            var result = _repository.Edit(order);
+
+            if (result == false)
+            {
+                return new ModificationResultDTO()
+                {
+                    Succeeded = false,
+                    Message = "Error updating the order"
+                };
+            }
+
+            return new ModificationResultDTO()
+            {
+                Succeeded = true
+            };
         }
 
-        public async Task<bool> DeleteObject(int orderId)
+        public async Task<ModificationResultDTO> DeleteObject(int orderId)
         {
             var order = await _repository.GetElement(x => x.Id == orderId);
+
+            if (order == null)
+            {
+                return new ModificationResultDTO()
+                {
+                    Succeeded = false,
+                    Message = "Order doesn't exist in the db"
+                };
+            }
+
             var result = _repository.Delete(order);
-            return result;
+
+            if (result == false)
+            {
+                return new ModificationResultDTO()
+                {
+                    Succeeded = false,
+                    Message = "Error deleting the order"
+                };
+            }
+
+            return new ModificationResultDTO()
+            {
+                Succeeded = true
+            };
         }
 
         public async Task<bool> SaveChangesForObject()
@@ -95,6 +159,77 @@ namespace Application.Services
             var orders = await _repository.GetPaginatedElements(pageNumber, pageSize); 
             var mappedOrders = _mapper.Map<List<DisplayOrderDTO>>(orders);
             return (mappedOrders, totalOrders);
+        }
+
+        public async Task<decimal> CalculateShipmentCost(Order order)
+        {
+            var settings = (await _settingsRepository.GetAllElements())[0];
+            var city = await _cityRepository.GetElement(c => c.id == order.CityId);
+            var merchant = await _merchantRepository.GetElement(m => m.Id == order.MerchantId, c => c.SpecialPackages);
+            var shipping = await _shippingRepository.GetElement(s => s.Id == order.ShippingId);
+            var orderType = order.Type;
+            var shippingType = shipping!.ShippingType;
+            var totalWeight = order.TotalWeight;
+            decimal shippingCost = 0;
+
+            switch (orderType)
+            {
+                case OrderTypes.BranchDelivery:
+
+                    var merchantSpecialPackage = merchant?.SpecialPackages.FirstOrDefault(sp => sp.cityId == order.CityId);
+                    
+                    if (merchantSpecialPackage != null)
+                    {
+                        shippingCost += merchantSpecialPackage.ShippingPrice;
+                    }
+                    else
+                    {
+                        shippingCost += city!.normalShippingCost;
+                    }
+
+                break;
+
+                case OrderTypes.HomeDelivery:
+
+                    if (merchant?.SpecialPickupShippingCost != null)
+                    {
+                        shippingCost += (decimal)merchant.SpecialPickupShippingCost;
+                    }
+                    else
+                    {
+                        shippingCost += city!.pickupShippingCost;
+                    }
+
+                break;
+            }
+
+            switch (shippingType)
+            {
+                case ShippingTypes.Ordinary:
+                    shippingCost += settings.OrdinaryShippingCost;
+                break;
+
+                case ShippingTypes.Within24Hours:
+                    shippingCost += settings.TwentyFourHoursShippingCost;
+                break;
+
+                case ShippingTypes.Within15Days:
+                    shippingCost += settings.FifteenDayShippingCost;
+                break;
+            }
+
+            if (totalWeight > settings.BaseWeight)
+            {
+                decimal additionalWeight = totalWeight - settings.BaseWeight;
+                shippingCost += additionalWeight * settings.AdditionalFeePerKg;
+            }
+
+            if (order.ShippingToVillage)
+            {
+                shippingCost += settings.VillageDeliveryFee;
+            }
+
+            return shippingCost;
         }
     }
 }
